[PLAT-12758] Add slow disks anomaly + bugfixes
Summary:
Implemented slow disks anomaly.
Also fixed few bugs, found by team running TF against real universe.

Test Plan:
Will write some unit tests later for slow disks + threshold exceed anomaly detection. Want to land fixes ASAP.
Also tested against real universes (local + dev portal one).
Seem to work.

Reviewers: cdavid, rmadhavan

Reviewed By: cdavid

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D33074

+++ troubleshoot/backend/src/main/java/com/yugabyte/troubleshoot/ts/service/TsStorageGraphService.java
@@ -255,7 +255,7 @@ public class TsStorageGraphService implements GraphSourceIF {
     TsStorageGraphConfig.FilterColumn column = config.getFilterColumns().get(filterName);
     return switch (column.getType()) {
       case type_text -> stringValues;
-      case type_int -> stringValues.stream().map(Integer::valueOf).toList();
+      case type_int -> stringValues.stream().map(Long::valueOf).toList();
       case type_float -> stringValues.stream().map(Double::valueOf).toList();
       case type_bool -> stringValues.stream().map(Boolean::valueOf).toList();
       case type_uuid -> stringValues.stream().map(UUID::fromString).toList();
